Updated Admin Page
Added functionality to edit usernames and superuser status from the admin page to make it easier to add moderators and edit usernames, if there are usernames that shouldn't be allowed. Also added placeholders for other possible future pages. Also added admin and user roles passed through the context of a response based on is_superuser.